moonbeam_rpc_core_debug

Trait DebugServer

pub trait DebugServer:
    Sized
    + Send
    + Sync
    + 'static {
    // Required methods
    fn trace_transaction<'life0, 'async_trait>(
        &'life0 self,
        transaction_hash: H256,
        params: Option<TraceParams>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionTrace>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_call<'life0, 'async_trait>(
        &'life0 self,
        call_params: TraceCallParams,
        id: RequestBlockId,
        params: Option<TraceParams>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionTrace>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn trace_block<'life0, 'async_trait>(
        &'life0 self,
        id: RequestBlockId,
        params: Option<TraceParams>,
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<BlockTransactionTrace>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self> { ... }
}

Server trait implementation for the Debug RPC API.

Required Methods

fn trace_transaction<'life0, 'async_trait>( &'life0 self, transaction_hash: H256, params: Option<TraceParams>, ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionTrace>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn trace_call<'life0, 'async_trait>( &'life0 self, call_params: TraceCallParams, id: RequestBlockId, params: Option<TraceParams>, ) -> Pin<Box<dyn Future<Output = RpcResult<TransactionTrace>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn trace_block<'life0, 'async_trait>( &'life0 self, id: RequestBlockId, params: Option<TraceParams>, ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<BlockTransactionTrace>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Provided Methods

fn into_rpc(self) -> RpcModule<Self>

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.